new try

+++ hashmap.c
@@ -40,7 +40,7 @@ int is_equal(void* key1, void* key2){
 
 
 void insertMap(HashMap * map, char * key, void * value) {
-	int i = hash(key, map->capacity);
+	unsigned long i = hash(key, map->capacity);
 
 	while(map->buckets[i] != NULL && map->buckets[i]->key != 0)
 	{
@@ -77,7 +77,7 @@ void eraseMap(HashMap * map,  char * key) {
 }
 
 Pair * searchMap(HashMap * map,  char * key) {   
-  int i = hash(key, map->capacity);
+  unsigned int i = hash(key, map->capacity);
 
   while(map->buckets[i] != NULL)
   {
